requirements.txt
asgiref==3.7.2
Django==4.2.5
django-crontab==0.7.1
django-jazzmin==2.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
icalendar==5.0.10
Pillow==10.0.1
psycopg2==2.9.7
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
whitenoise==6.6.0
